msm: mdss: calculate mdp ib bandwidth with overlap and smp factor
mdp ib request for multiple small layer use cases is too large because
currently it is the lump sum of all individual layer requests times a
constantant factor. Adding the logic to calculate the worst case of ib
requests for overlapped areas and the concurrent smp fetching for all
layers.

Signed-off-by: Huaibin Yang <huaibiny@codeaurora.org>
Signed-off-by: Naseer Ahmed <naseer@codeaurora.org>